requirements.txt
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0